Extracted interaction menu logic
