+++ planet/Objects.ocd/Clonk.ocd/InteractionMenu.ocd/StringTblDE.txt
@@ -1,0 +1,1 @@
+MsgDeadClonk=%s ist tot.|Es sind keine Interaktionen möglich.
