+++ planet/Objects.ocd/Clonk.ocd/StringTblDE.txt
@@ -8,5 +8,3 @@ CtxChop=Fällen
 CtrlMenuDesc=Menü
 TxtDescend=Abspringen
 TxtEnergysupply=Energieversorgung
-
-MsgDeadClonk=%s ist tot.|Es sind keine Interaktionen möglich.
