+++ planet/Objects.ocd/Clonk.ocd/StringTblUS.txt
@@ -8,5 +8,3 @@ CtxChop=Chop
 CtrlMenuDesc=Menu
 TxtDescend=Jump off
 TxtEnergysupply=Energy supply
-
-MsgDeadClonk=%s is dead.|No interactions are possible.
